Cleaned up a small warning on compile. Added an option to show elapsed time it took the mud to initialize.

+++ trunk/src/socials.cpp
@@ -117,7 +117,7 @@ void Socials::Load(void)
           data->rtarg = social->Attribute("rtarg");
           data->ytarg = social->Attribute("ytarg");
 
-          if (id > _socid)
+          if ((unsigned int)id > _socid)
             {
               _socid = id+1;
             }
